refactor: convert src/ to consistent package imports
Switches src/ from a mix of top-level absolute imports and try/except
hybrids to consistent relative package imports. This fixes the regression
where `python -m src.cli` failed without PYTHONPATH=src — the documented
CLI invocation in README.md now works directly from a fresh checkout.

- pyproject.toml: pytest pythonpath src -> .
- src/build_feed.py, src/cli.py, src/feed/{config,logging,logging_safe,
  providers,reporting}.py, src/places/{client,merge,quota}.py,
  src/providers/wl_fetch.py, src/utils/env.py: drop script-mode fallbacks
- src/utils/env.py: drop 95-line duplicate sanitize_log_message fallback
  (always reachable via .logging now that src is a real package)
- .github/workflows/build-feed.yml: use python -m src.cli feed build
- tests: switch 3 top-level imports to src.X; stop popping src.providers
  from sys.modules in test_build_feed_cache helper (caused pollution of
  later wl_fetch/vor tests); remove obsolete fallback tests

Verified: python -m src.cli --help and feed build run without PYTHONPATH;
990 tests pass (6 fallback tests removed as their target code is gone).
