+++ src/feed/logging.py
@@ -16,11 +16,7 @@ from .config import (
     LOG_TIMEZONE,
 )
 
-# Import the new safe formatters
-try:
-    from .logging_safe import SafeFormatter, SafeJSONFormatter, _make_formatter
-except ImportError:
-    from feed.logging_safe import SafeFormatter, SafeJSONFormatter, _make_formatter  # type: ignore[no-redef]
+from .logging_safe import SafeFormatter, SafeJSONFormatter, _make_formatter
 
 LOG_DIR = LOG_DIR_PATH.as_posix()
 error_log_path = Path(LOG_DIR) / "errors.log"
