+++ src/utils/env.py
@@ -24,99 +24,9 @@ import logging
 import os
 import re
 from pathlib import Path
-from typing import Dict, Iterable, List, Mapping, MutableMapping
+from typing import Dict, Iterable, Mapping, MutableMapping
 
-try:
-    from .logging import sanitize_log_message
-except ImportError:
-    try:
-        from utils.logging import sanitize_log_message  # type: ignore[no-redef]
-    except ImportError:
-        # Fallback security masker if logging module is unreachable.
-        # This ensures secrets are not leaked in plaintext during import errors.
-
-        # Precompiled regexes for sanitization (copied from src.utils.logging)
-        _CONTROL_CHARS_RE = re.compile(r"[\x00-\x1f\x7f-\x9f]")
-        # ANSI escape codes: comprehensive matching for CSI, OSC, Fe, and 2-byte sequences
-        # Matches:
-        # 1. CSI: ESC [ ...
-        # 2. OSC: ESC ] ... BEL/ST
-        # 3. Fe (excluding [ and ]): ESC [@-Z\\^_]
-        # 4. Two-byte sequences: ESC [space-/] [0-~]
-        _ANSI_ESCAPE_RE = re.compile(r'\x1b(?:\[[0-?]*[ -/]*[@-~]|\][^\x07\x1b]*(?:\x07|\x1b\\)|[@-Z\\^_]|[\x20-\x2f][\x30-\x7e])')
-
-        def sanitize_log_message(
-            text: str, secrets: List[str] | None = None, strip_control_chars: bool = True
-        ) -> str:
-            if not text:
-                return ""
-            sanitized = text
-
-            # Remove ANSI escape codes explicitly first
-            sanitized = _ANSI_ESCAPE_RE.sub("", sanitized)
-
-            # Comprehensive keys list mirroring src.utils.logging to ensure safety during fallback
-            _keys = (
-                r"client[-_.\s]*secret|access[-_.\s]*token|refresh[-_.\s]*token|[a-z0-9_.\-]*client[-_.\s]*id[a-z0-9_.\-]*|[a-z0-9_.\-]*signature|[a-z0-9_.\-]*password[a-z0-9_.\-]*|[a-z0-9_.\-]*e[-_.\s]*mail[a-z0-9_.\-]*|"
-                r"client[-_.\s]*assertion[-_.\s]*type|client[-_.\s]*assertion|"
-                r"saml[-_.\s]*request|saml[-_.\s]*response|"
-                r"[a-z0-9_.\-]*accessid[a-z0-9_.\-]*|id[-_.\s]*token|[a-z0-9_.\-]*session[-_.\s]*id[a-z0-9_.\-]*|session|cookie|[a-z0-9_.\-]*apikey[a-z0-9_.\-]*|[a-z0-9_.\-]*secret[a-z0-9_.\-]*|ticket|[a-z0-9_.\-]*token|code|key|sig|sid|"
-                r"nonce|state|"
-                r"jsessionid|phpsessid|asp\.net_sessionid|__cfduid|"
-                r"authorization|auth|bearer[-_.\s]*token|bearer|[a-z0-9_.\-]*api[-_.\s]*key[a-z0-9_.\-]*|[a-z0-9_.\-]*private[-_.\s]*key|auth[-_.\s]*token|"
-                r"tenant[-_.\s]*id|tenant|subscription[-_.\s]*id|subscription|object[-_.\s]*id|oid|"
-                r"code[-_.\s]*challenge|code[-_.\s]*verifier|"
-                r"x[-_.\s]*api[-_.\s]*key|ocp[-_.\s]*apim[-_.\s]*subscription[-_.\s]*key|"
-                r"[a-z0-9_.\-]*credential|x[-_.\s]*amz[-_.\s]*credential|x[-_.\s]*amz[-_.\s]*security[-_.\s]*token|"
-                r"x[-_.\s]*amz[-_.\s]*signature|x[-_.\s]*auth[-_.\s]*token|"
-                r"[a-z0-9_.\-]*passphrase[a-z0-9_.\-]*|[a-z0-9_.\-]*access[-_.\s]*key[-_.\s]*id[a-z0-9_.\-]*|"
-                r"[a-z0-9_.\-]*secret[-_.\s]*access[-_.\s]*key|[a-z0-9_.\-]*auth[-_.\s]*code[a-z0-9_.\-]*|"
-                r"[a-z0-9_.\-]*authorization[-_.\s]*code[a-z0-9_.\-]*|"
-                r"[a-z0-9_.\-]*otp(?:[-_][a-z0-9_.\-]*)?|[a-z0-9_.\-]*glpat(?:[-_][a-z0-9_.\-]*)?|[a-z0-9_.\-]*ghp(?:[-_][a-z0-9_.\-]*)?|"
-                r"\bpass\b|\bpwd\b|\buser[-_.]?pass\b"
-            )
-
-            _header_keys = (
-                r"api[-_.\s]*key|token|secret|signature|password|auth|session|cookie|private|"
-                r"client[-_.\s]*assertion|saml[-_.\s]*request|saml[-_.\s]*response|nonce|state|"
-                r"credential|client[-_.\s]*id|passphrase|access[-_.\s]*key|e[-_.\s]*mail"
-            )
-
-            # Simplified patterns for fallback (subset of full logging module but covering critical cases)
-            patterns = [
-                # PEM blocks (keys/certs)
-                (r"(-----BEGIN [A-Z ]+-----)(?:.|\n)*?(-----END [A-Z ]+-----)", r"\1***\2"),
-                # URL credentials
-                (r"(?i)([a-z0-9+.-]+://)([^/@\s]+)@", r"\1***@"),
-                # Query params and assignments (key=value)
-                (
-                    rf"(?i)((?:{_keys})\s*(?:%3d|=)\s*)"
-                    rf"((?:\"(?:\\.|[^\"\\\\])*\")|(?:'(?:\\.|[^'\\\\])*')|((?:(?!\s+[a-zA-Z0-9_.-]+\s*(?:%3d|=))[^&,\n])+))",
-                    r"\1***",
-                ),
-                # JSON fields (key: "value")
-                (rf'(?i)(\"(?:{_keys})\"\s*:\s*\")((?:\\.|[^"\\\\])*)(\")', r'\1***\3'),
-                (rf"(?i)('(?:{_keys})'\s*:\s*')((?:\\.|[^'\\\\])*)(')", r"\1***\3"),
-                # Headers
-                (rf"(?i)((?:[-a-zA-Z0-9_]*(?:{_header_keys})[-a-zA-Z0-9_]*):\s*)((?:.*)(?:\n\s+.*)*)", r"\1***"),
-            ]
-
-            for pattern, repl in patterns:
-                sanitized = re.sub(pattern, repl, sanitized)
-
-            # Mask secrets explicitly provided
-            if secrets:
-                for secret in secrets:
-                    if secret:
-                        sanitized = sanitized.replace(secret, "***")
-
-            # Escape control characters to prevent log injection
-            if strip_control_chars:
-                # We escape common control chars to keep the log readable but safe
-                sanitized = sanitized.replace("\n", "\\n").replace("\r", "\\r").replace("\t", "\\t")
-                sanitized = _CONTROL_CHARS_RE.sub("", sanitized)
-
-            return sanitized
+from .logging import sanitize_log_message
 
 __all__ = [
     "get_int_env",
